constrain the change to just the plugin
